FEAT :: ReservationLog entity 구현
